refactor: Move ScenarioError into smite
Despite the error name `ScenarioError`, `Scenario::new` was expecting `String`
as errors. This also meant we had to map `ScenarioError` to `String` with
`map_err` in a lot of places.

This moves `ScenarioError` to the smite package, and uses it instead of `String`.

To avoid a circular dependency between smite-scenarios and smite, `TargetError`
is also moved.
