+++ smite-scenarios/Cargo.toml
@@ -23,4 +23,3 @@ libc = "0.2"
 serde.workspace = true
 serde_json = "1"
 tempfile = "3"
-thiserror = "2"
